+++ metalsapi/test/adapter_test.js
@@ -19,6 +19,24 @@ describe('execute', () => {
         name: 'from/to',
         testData: { id: jobID, data: { from: 'CHF', to: 'USD' } },
       },
+      {
+        name: 'coin',
+        testData: { id: jobID, data: { coin: 'BTC' } },
+      },
+      {
+        name: 'BTC difficulty',
+        testData: {
+          id: jobID,
+          data: { blockchain: 'BTC' },
+        },
+      },
+      {
+        name: 'BTC height',
+        testData: {
+          id: jobID,
+          data: { blockchain: 'BTC', endpoint: 'height' },
+        },
+      },
     ]
 
     requests.forEach((req) => {